requirements.txt
request
bs4
pandas
googletrans
numpy
